% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\alias{fraction}
\title{Safe division of numeric vectors}
\usage{
divide(a, b)

fraction(a, b)
}
\arguments{
\item{a}{Numerator}

\item{b}{Denominator}
}
\value{
A numeric vector
}
\description{
Safe division of numeric vectors
}
\examples{
multiply(1, 2)
multiply(1:10, rep(2, 10))
}
