% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{division_table}
\alias{division_table}
\title{Division table}
\format{
A tidy data frame with 100 rows and 3 columns:
\describe{
\item{x}{Numerator}
\item{y}{Denominator}
\item{z}{Quotient}
}
}
\usage{
division_table
}
\description{
A long tidy tibble containing a division table
}
\keyword{datasets}
