% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%*/\%}
\alias{\%*/\%}
\title{Either multiplication or division}
\usage{
LHS \%*/\% RHS
}
\arguments{
\item{LHS, RHS}{Arguments to either multiply or divide}
}
\value{
A numeric vector
}
\description{
Either multiplication or division
}
\examples{
4 \%*/\% 2
1:10 \%*/\% 11:20
}
