% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiplication_table}
\alias{multiplication_table}
\title{Multiplication table}
\format{
A tidy data frame with 100 rows and 3 columns:
\describe{
\item{x}{Multiplicand}
\item{y}{Multiplier}
\item{z}{Product}
}
}
\usage{
multiplication_table
}
\description{
A long tidy tibble containing a multiplication table
}
\keyword{datasets}
