% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Safe multiplication of numeric vectors}
\usage{
multiply(a, b)
}
\arguments{
\item{a}{First element to be multiplied}

\item{b}{Second element to be multiplied}
}
\value{
A numeric vector
}
\description{
Safe multiplication of numeric vectors
}
\examples{
multiply(1, 2)
multiply(1:10, rep(2, 10))
}
